+++ shoop/admin/modules/orders/static_src/create/reducers/index.js
@@ -12,11 +12,22 @@ import productData from "./productData";
 import shop from "./shop";
 import customer from "./customer";
 import customerData from "./customerData";
+import customerDetails from "./customerDetails";
 import methods from "./methods";
 import order from "./order";
 import comment from "./comment";
 
-const childReducer = combineReducers({lines, productData, shop, customer, customerData, methods, order, comment});
+const childReducer = combineReducers({
+    lines,
+    productData,
+    shop,
+    customer,
+    customerData,
+    customerDetails,
+    methods,
+    order,
+    comment
+});
 
 export default function(state, action) {
     if(action.type === "_replaceState") { // For debugging purposes.