+++ shoop/admin/modules/orders/static_src/create/view/utils.js
@@ -40,14 +40,63 @@ export function selectBox(value, onchange, choices, valueGetter = "id", nameGett
     ));
 }
 
-export function contentBlock(icon, title, view) {
+export function contentBlock(icon, title, view, header = "h2") {
     return m("div.content-block",
         m("div.title",
-            m("h2.block-title", m(icon), " " + title),
+            m(header + ".block-title", m(icon), " " + title),
             m("a.toggle-contents", m("i.fa.fa-chevron-right"))
         ),
         m("div.content-wrap.collapse",
             m("div.content", view)
+        )
+    );
+}
+
+export function infoRow(header, value, klass) {
+    if(value && value !== ""){
+        return m("div", [
+            m("dt", header),
+            m("dd" + (klass? klass : ""), value)
+        ]);
+    }
+}
+
+export function table({columns, data=[], tableClass="", emptyMsg=gettext("No records found.")}) {
+    return m("table", {class: "table " + tableClass},
+        m("thead",
+            m("tr", columns.map(col => m("th", col.label)))
+        ),
+        m("tbody",
+            (data.length > 0?
+            data.map(row => m("tr", columns.map(col => m("td", row[col.key])))):
+            m("tr", m("td[colspan='" + columns.length + "'].text-center", m("i", emptyMsg))))
+        )
+    );
+}
+
+export function modal({show=false, sizeClass="", title, body, footer, close}) {
+    if(show){
+        $("body").append("<div class='modal-backdrop in'></div>").addClass("modal-open");
+    } else {
+        $("body").removeClass("modal-open").find(".modal-backdrop").remove();
+    }
+
+    return (
+        m("div.modal" + (show? " show": " hidden"),
+            m("div.modal-dialog", {class: sizeClass},
+                m("div.modal-content",
+                    m("div.modal-header",
+                        m("button.close",
+                            m("span", {
+                                onclick: () => close()
+                            }, m.trust("&times;"))
+                        ),
+                        title
+                    ),
+                    m("div.modal-body", body),
+                    m("div.modal-footer", footer)
+                )
+            )
         )
     );
 }